+++ src/engine/physics/physics_engine.h
@@ -28,7 +28,7 @@ public:
     void unregisterComponent(engine::component::PhysicsComponent* component); ///< @brief 注销物理组件
 
     ///< @brief 核心循环：更新所有注册的物理组件的状态
-    void update(double deltaTime);
+    void update(float deltaTime);
 
     // 设置器/获取器
     ///< @brief 设置全局重力加速度
